服务端添加机器人事件

+++ main/java/com/robot/demo/domain/RobotAngleData.java
@@ -4,11 +4,11 @@ import lombok.Data;
 
 @Data
 public class RobotAngleData {
-    private int angleValue = 0;
+    private double angleValue = 0;
     private int angleNo = 0;
     private int robotNo = 0;
 
-    public RobotAngleData(int angleNo, int angleValue, int robotNo) {
+    public RobotAngleData(int angleNo, double angleValue, int robotNo) {
         this.angleValue = angleValue;
         this.angleNo = angleNo;
         this.robotNo = robotNo;
